requirements.txt
PyOpenGL==3.1.5
numpy==1.21.4
moderngl==5.6.4
moderngl_window==2.4.0
pyrr==0.10.3
QScintilla==2.13.1
pillow==9.0.0
